+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 🚀 Features
 
-- **web**: Add sample deployment webpage - by @thuongtruong109 [<samp>(46c26)</samp>](https://github.com/thuongtruong109/flashot/commit/46c26cdf4127e5b2a01e5bc48945975e9c3aff18)
+- **docs**: add landing page - by @thuongtruong109 [<samp>(33c8e)</samp>](https://github.com/thuongtruong109/flashot/commit/33c8e77580cbff104125642ebe7fc55be2253963)
+- **web**: Add sample playground - by @thuongtruong109 [<samp>(46c26)</samp>](https://github.com/thuongtruong109/flashot/commit/46c26cdf4127e5b2a01e5bc48945975e9c3aff18)
 - **refactor**: Refactor codebase to workspace monorepo - by @thuongtruong109 [<samp>(c019d)</samp>](https://github.com/thuongtruong109/flashot/commit/c019d61bfd82c01293563139ba13d5e59a2f443f)
 
 ### 🐞 Bug Fixes